:recycle: refactor file structure
