+++ backend/src/application/session_manager.py
